requirements_tests.txt
django-jenkins
pep8
pyflakes
clonedigger
flake8
flake8-todo
coverage
